feat: combineClassNames util 함수 생성
